Extract BankSetter
