+++ src/commonMain/kotlin/choliver/nespot/cartridge/ChrMemory.kt
@@ -9,8 +9,17 @@ class ChrMemory(
   private val bankSize: Int = raw.size,
   var mirroring: Mirroring = IGNORED
 ) : Memory {
-  private val vram = ByteArray(VRAM_SIZE)
+  init {
+    if ((bankSize % PAGE_SIZE) != 0) {
+      throw IllegalArgumentException("Bank size ${bankSize} not a multiple of page size ${PAGE_SIZE}")
+    }
+  }
+
+  // We map at page granularity, rather than bank granularity
+  private val pagesPerBank = bankSize / PAGE_SIZE
+  private val pageMap = IntArray(NUM_PAGES) { it * PAGE_SIZE }
   val bankMap = IntArray(CHR_SIZE / bankSize) { it }
+  private val vram = ByteArray(VRAM_SIZE)
 
   override fun get(addr: Address) = when {
     (addr >= BASE_VRAM) -> vram[vmap(addr)]    // This maps everything >= 0x4000 too
@@ -32,5 +41,10 @@ class ChrMemory(
     VERTICAL -> addr and 2047
     HORIZONTAL -> (addr and 1023) or ((addr and 2048) shr 1)
     IGNORED -> throw UnsupportedOperationException()
+  }
+
+  companion object {
+    private const val PAGE_SIZE = 1024    // This should be no bigger than the minimum known bank size
+    private const val NUM_PAGES = CHR_SIZE / PAGE_SIZE
   }
 }
